pyproject.toml
[project]
name = "ezbookkeeping-mcp-server"
version = "0.1.0"
description = "MCP server for EzBookkeeping personal finance management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "mcp[cli]>=1.21.2",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/theanirudhvyas/ezbookkeeping_mcp_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
